255-byte receive code integrated
Solved the problems associated with receiving the 255-byte packets.
CC1101 radio now automatically defaults to RX mode when finished
receiving  or transmitting a packet to ensure it will be ready to
receive a new sync word. In IDLE mode the radio is unable to receive new
sync words and therefore will not generate the required interrupts.

One issue that has not be resolved yet is that in order to receive
multiple packets without resetting the radio, it is necessary to
manually specify the packet length in code. Reading the packet length
from the first data byte only works for the first packet recieve and
after that there is a shift where the first data packet no longer is the
length and this messes up the code. Why this is happening is under
investigation.