Working TX Code
Fixed the length byte to reflect the payload data EXLCUDING itself. This
solved the extra packet error I was having as well as the TX UNDERFLOW
error. In order to deal with the TX UNDERFLOW I built a check in the
code for that so in case it does happen at some point the program will
still be able to roll on.

Also, I have made the functionality more robust so the code can transmit
any number of bytes from 1 to 255. Previously the control flow would not
send any packet length < the TX threshold (currently set at 33).

The RX code still needs  work in order to have consistent functionality.